+++ permissions/src/epilot/assignments.py
@@ -68,7 +68,7 @@ class Assignments:
         
         url = utils.generate_url(operations.AssignRolesRequest, base_url, '/v1/permissions/assignments/{userId}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "request_body", False, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.AssignRolesRequest, "request_body", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'